+++ constructors/CHANGELOG.md
@@ -4,8 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - 2021-04-23
+## [1.0.0] - 2021-04-28
 
 ### Added
 * Initial implementation of the module.
-* Add buffer function based on turfjs v6.3.0.
+* Add ST_BUFFER function based on turfjs v6.3.0.
+* Add ST_MAKEENVELOPE function
+* Add ST_TILEENVELOPE function
+